+++ ImmersionRC.LapRF/Communication/LapRFCommunicationFactory.cs
@@ -4,5 +4,5 @@ using System.Net;
 namespace ImmersionRC.LapRF.Communication;
 internal class LapRFCommunicationFactory : ILapRFCommunicationFactory
 {
-    public ILapRFCommunication CreateEthernetCommunication(IPAddress address, int port) => new LapRFEthernetCommunication(address, port);
+    public ILapRFCommunication CreateEthernetCommunication(IPAddress address, ushort port) => new LapRFEthernetCommunication(address, port);
 }
