+++ NaeTime.Hardware.ImmersionRC/Abstractions/ILapRFConnectionFactory.cs
@@ -3,5 +3,5 @@
 namespace NaeTime.Timing.ImmersionRC.Abstractions;
 internal interface ILapRFConnectionFactory
 {
-    LapRFConnection CreateEthernetConnection(Guid timerId, IPAddress address, int port);
+    LapRFConnection CreateEthernetConnection(Guid timerId, IPAddress address, ushort port);
 }
